Refactor things out of utils.h
